evaluation-examen added to user controller + some fixes

+++ src/main/java/com/ecole/serviceimpl/MatiereServiceImpl.java
@@ -1,6 +1,7 @@
 package com.ecole.serviceimpl;
 
 import com.ecole.domain.Matiere;
+import com.ecole.domain.Question;
 import com.ecole.repository.MatiereRepository;
 import com.ecole.service.MatiereService;
 import lombok.RequiredArgsConstructor;
@@ -20,28 +21,42 @@ public class MatiereServiceImpl implements MatiereService {
     private MatiereRepository matiereRepository;
     @Override
     public Matiere addMatiere(Matiere matiere) {
+        log.info("Enregistrement nouvelle Matiere {} à la BD",matiere.getTitle());
         return this.matiereRepository.save(matiere);
     }
 
     @Override
     public Matiere updateMatiere(Matiere matiere) {
+        log.info("Mise à jour Matiere {} ",matiere.getTitle());
         return this.matiereRepository.save(matiere);
     }
 
     @Override
     public Set<Matiere> getMatieres() {
+        log.info("Toutes les matieres");
         return new LinkedHashSet<>(this.matiereRepository.findAll());
 
     }
 
     @Override
     public void deleteMatiere(Long idmat) throws Exception {
+        Matiere matiere=this.matiereRepository.findById(idmat).get();
+        if(matiere==null) {
+            log.error("Id de la matiere invalide");
+            throw new Exception("Id de la matiere invalide");
+        }
+        log.info("Suppression de la matiere {}",matiere.getTitle());
         this.matiereRepository.deleteById(idmat);
     }
 
     @Override
     public Matiere getMatiereById(Long idmat) throws Exception {
         Matiere matiere=this.matiereRepository.findById(idmat).get();
+        if(matiere==null) {
+            log.error("Id de la matiere invalide");
+            throw new Exception("Matiere inconnue");
+        }
+        log.info("Extraction matiere ");
         System.out.println("Matiere:"+matiere.getTitle()+" "+matiere.getDescription()+" "+matiere.getIdMat());
         return matiere;
     }
